requirements.txt
gradio>=3.0
python-docx>=0.8.11
openai>=0.27.0
faiss-cpu
sentence-transformers
pydantic
